Revert "scintilla: Updated contrib to 3.6.0."
This reverts commit 25004787bd2f713aada7d0ef398c4552ba9f5a5f.

+++ contrib/scintilla/cocoa/QuartzTextLayout.h
@@ -85,7 +85,7 @@ public:
 		if (mLine == NULL)
 			return 0.0f;
 
-		return static_cast<float>(::CTLineGetTypographicBounds(mLine, NULL, NULL, NULL));
+		return ::CTLineGetTypographicBounds(mLine, NULL, NULL, NULL);
 	}
 
     CTLineRef getCTLine() {
